changing scroll to intersect and adding modifiers to intersect directive
