+++ src/directives/index.js
@@ -1,4 +1,4 @@
 export { default as autofocus } from "./autofocus"
 export { default as clickout } from "./clickout"
 export { default as copy } from "./copy"
-export { default as scroll } from './scroll'
+export { default as intersect } from './intersect'
